move: flyingstar 패키지 이동
