+++ on-fe/app/(auth)/login/additional-info/page.js
@@ -2,7 +2,7 @@
 import AdditionalInfoForm from '@/components/auth/AdditionalInfoForm';
 import FormLayout from '@/components/form/FormLayOut';
 import '@/app/(auth)/login/additional-info/info-form.css'
-import FlyingStar from '@/app/ui/background/FlyingStar';
+import FlyingStar from '@/components/background/FlyingStar';
 
 export default function AdditionalInfoPage() {
     return (
